+++ CHANGELOG.rst
@@ -31,6 +31,8 @@ Added
 - add graph modification when less than graph returns are specified
 - add "__name__" attribute to handler instance
 - ``signature_modifier`` can modify function with "kwargs"
+- ``pos_signature_modifier`` allows for node object to have positional only parameters
+- graph node definition allows for builtin and numpy ufunc
 
 [0.4.0] - 2022-10-3
 ------------------------